functions/kits/requirements.txt
boto3
flask-lambda-support
pyhumps